+++ include/soloud_dcremovalfilter.hpp
@@ -40,9 +40,9 @@ class DCRemovalFilterInstance : public FilterInstance
 
   public:
     void filter(float*       aBuffer,
-                unsigned int aSamples,
-                unsigned int aBufferSize,
-                unsigned int aChannels,
+                size_t aSamples,
+                size_t aBufferSize,
+                size_t aChannels,
                 float        aSamplerate,
                 time_t       aTime) override;
     ~DCRemovalFilterInstance() override;
